Added gender choice, date, and email tests
The gender choice test checks if the gender choice data is valid. The
options for now are m for male, f for female, o for other, and d for
do not want to specify.

The date tests check if the date data is an integer and if it is a
valid date.

The email test checks if the email data is a valid email string. It
does not check if that email actually exists.