Updated RegisteredUser Model
The RegisteredUser model is updated to accomodate the class diagram.
The model is updated to include an additional field named account.
Account field is a foreign key to the Account model. The RegisteredUser
model also has two new boolean settings fields which are:
- is_messaging_allowed
- is_notification_allowed

The Account model holds the following optional information fields:
- firstname
- lastname
- phone_number
- verified_as_doctor
- profile_picture
- profession
- location
- diplomaID