Profile Edit Functionality Implemented And Fix to the GET Method
Edit Profile Functionality

The profile endpoint now allows the user to edit their profile by
sending a PATCH request to the endpoint. The user can edit their
- email
- gender
- birth_date
- first_name
- last_name
- phone_number
- profile_picture
- profession
- location
- diplomaID

The users can also remove any information above by sending an empty
string with the field name as key in the request body form data.

Fix to the GET Method

The get method now returns the information in the account of a user
correctly.