Comment Model Added
The model class for the comment table has been added to the models.py
file. The class has the following attributes:

- parent_comment
- parent_post

Note that the parent_comment and parent_post attributes are foreign
keys that can be blank or null. Only one of the two attributes can be
set at a time. This is because a comment can be a reply to a post or a
reply to another comment but not both at the same time.